+++ aofl-js-packages/map-state-properties-mixin/test/index.spec.js
@@ -1,5 +1,5 @@
 /* eslint-disable */
-import {Store} from '@aofl/store';
+import {Store} from '@aofl/store/modules/legacy';
 import {AoflElement} from '@aofl/web-components/aofl-element';
 import mapStatePropertiesMixin from '../modules/map-state-properties-mixin';
 import {render, html} from 'lit-html';