+++ aofl-js-packages/store/modules/legacy/instance.js
@@ -4,7 +4,7 @@
  *
  */
 
-import {Store} from '../../';
+import Store from './store';
 
 /**
  * Persistent instance of Store.